Subscriber/requirements.txt
Flask
paho-mqtt
xmltodict
requests
urllib3
